lastPlayed

+++ src/main/java/com/lgs/eden/api/nexus/ProfileImp.java
@@ -76,7 +76,7 @@ public class ProfileImp extends ImpSocket implements ProfileAPI {
                         g.getString("icon"),
                         g.getString("name"),
                         g.getInt("time_played") / 3600,
-                        g.getInt("last_played")
+                        g.getLong("last_played")
                 );
             }
